requirements.txt
fastapi
uvicorn
python-telegram-bot[asyncio]
